+++ src/utils/mod.rs
@@ -11,12 +11,12 @@ pub use tile_error::TileError;
 
 pub const TILE_SIZE_IN_PX: u32 = 256;
 
-pub fn calc_max_lod(image: &DynamicImage) -> u8 {
+pub fn calc_max_lod(image: &DynamicImage) -> usize {
     let width = image.dimensions().0 as f32;
 
     let tiles_per_row = (width / TILE_SIZE_IN_PX as f32).ceil();
 
-    return tiles_per_row.log2().ceil() as u8;
+    return tiles_per_row.log2().ceil() as usize;
 }
 
 pub fn encode_png(
